[external-assets] Move InternalAssetGraph to its own module
